requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-telegram-bot==13.7
langchain==0.3.0
numpy==1.21.4
faiss-cpu==1.7.1
python-dotenv==0.19.1
